New feature: Create security descriptors automatically based on the category of an article
http://dmwiki.ontoprise.com:8888/dmwiki/index.php/Create_security_descriptors_automatically_based_on_the_category_of_an_article


git-svn-id: http://halodev.ontoprise.com/svn/HaloSMWExtension/trunk@9599 f80b7aa0-dfcc-4e65-a0d9-74108174cbc8

+++ extensions/HaloACL/includes/HACL_GlobalFunctions.php
@@ -585,7 +585,7 @@ function haclfGetUserID($user = null) {
                     } else {
                     // name of user given
                         $etc = haclfDisableTitlePatch();
-                        $userID = User::idFromName($user);
+                        $userID = (int) User::idFromName($user);
                         haclfRestoreTitlePatch($etc);
                         if (!$userID) {
                             $userID = false;
